Update PUDialogueBox.cpp
Added lots of logging and also fixed the input issue where the mouse and control wasn't being given back to player. Also for now its a workaround and need to figure out a better way instead of the fallback of getting it from the world. Need to figure out how keep the player controller reference.