add in gameplay tags and refactoring data for ingredients and dishes
Hopefully will give us a more robust system to start building out a dish and creating the radar graphs. This is the first step. I want to go over later the flavor profiles and texture profiles so we can nail down the properties. That way we can also start tweaking them.